+++ packages/react-core/src/components/Menu/__mocks__/MenuItem.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { MenuItemProps } from '../MenuItem';
+
+export const MenuItem = ({ className, children, description, itemId }: MenuItemProps) => (
+  <>
+    <div className={className} data-testid="menu-item-mock">
+      {children}
+    </div>
+    <p>{`description: ${description}`}</p>
+    <p>{`itemId: ${itemId}`}</p>
+  </>
+);